api service 가독성 개선
