+++ app/src/main/java/com/example/wafflestudio_toyproject/VoteDetailViewModel.kt
@@ -4,9 +4,9 @@ import androidx.lifecycle.LiveData
 import androidx.lifecycle.MutableLiveData
 import androidx.lifecycle.ViewModel
 import androidx.lifecycle.map
-import androidx.lifecycle.viewModelScope
+import com.example.wafflestudio_toyproject.network.ParticipationRequest
+import com.example.wafflestudio_toyproject.network.VoteDetailResponse
 import dagger.hilt.android.lifecycle.HiltViewModel
-import kotlinx.coroutines.launch
 import javax.inject.Inject
 
 @HiltViewModel
