+++ app/src/main/java/com/example/wafflestudio_toyproject/adapter/VoteItemAdapter.kt
@@ -2,11 +2,10 @@ package com.example.wafflestudio_toyproject.adapter
 
 import android.view.LayoutInflater
 import android.view.ViewGroup
-import androidx.recyclerview.widget.DiffUtil
 import androidx.recyclerview.widget.RecyclerView
 import coil.load
 import com.example.wafflestudio_toyproject.R
-import com.example.wafflestudio_toyproject.VoteItem
+import com.example.wafflestudio_toyproject.network.VoteItem
 import com.example.wafflestudio_toyproject.databinding.VoteItemBinding
 
 class VoteItemAdapter(
