+++ app/src/main/java/com/example/wafflestudio_toyproject/fragments/UserPage/UserProfileFragment.kt
@@ -11,10 +11,10 @@ import android.widget.Toast
 import androidx.fragment.app.Fragment
 import androidx.navigation.NavController
 import androidx.navigation.fragment.findNavController
-import com.example.wafflestudio_toyproject.GetMeResponse
+import com.example.wafflestudio_toyproject.network.GetMeResponse
 import com.example.wafflestudio_toyproject.LoginActivity
 import com.example.wafflestudio_toyproject.R
-import com.example.wafflestudio_toyproject.UserApi
+import com.example.wafflestudio_toyproject.network.UserApi
 import com.example.wafflestudio_toyproject.UserRepository
 import com.example.wafflestudio_toyproject.databinding.FragmentUserProfileBinding
 import dagger.hilt.android.AndroidEntryPoint
@@ -88,8 +88,6 @@ class UserProfileFragment : Fragment() {
     
     // 회원 정보 불러오기
     private fun getUserInformation() {
-        val accessToken = "Bearer ${userRepository.getAccessToken()}"
-
         val colleges = listOf(
             "인문대학", "사회과학대학", "자연과학대학", "간호대학", "경영대학", "공과대학",
             "농업생명과학대학", "미술대학", "사범대학", "생활과학대학", "수의과대학", "약학대학", "음악대학", "의과대학",
@@ -97,7 +95,7 @@ class UserProfileFragment : Fragment() {
         )
 
         // getMe API 호출
-        userApi.getMe(accessToken).enqueue(object : Callback<GetMeResponse> {
+        userApi.getMe().enqueue(object : Callback<GetMeResponse> {
             override fun onResponse(call: Call<GetMeResponse>, response: Response<GetMeResponse>) {
                 if (response.isSuccessful) {
                     response.body()?.let { userInfo ->
