+++ app/src/main/java/com/example/wafflestudio_toyproject/network/ApiModels.kt
@@ -1,105 +1,8 @@
-package com.example.wafflestudio_toyproject
+package com.example.wafflestudio_toyproject.network
 
-import okhttp3.MultipartBody
-import okhttp3.RequestBody
-import okhttp3.ResponseBody
-import retrofit2.Call
-import retrofit2.http.Body
-import retrofit2.http.DELETE
-import retrofit2.http.GET
-import retrofit2.http.Header
-import retrofit2.http.Multipart
-import retrofit2.http.PATCH
-import retrofit2.http.POST
-import retrofit2.http.Part
-import retrofit2.http.Path
-import retrofit2.http.Query
 import java.text.SimpleDateFormat
 import java.util.Date
 import java.util.Locale
-
-interface UserApi {
-    @POST("api/users/signup")
-    fun signup(@Body signupRequest: SignupRequest): Call<SignupResponse>
-
-    @POST("api/users/signin")
-    fun login(@Body loginRequest: LoginRequest): Call<LoginResponse>
-
-    @GET("api/users/me")
-    fun getMe(
-        @Header("Authorization") authorization: String
-    ): Call<GetMeResponse>
-
-    @PATCH("api/users/reset_pw")
-    fun changePassword(
-        @Header("Authorization") authorization: String,
-        @Body changePasswordRequest: ChangePasswordRequest
-    ): Call<ResponseBody>
-}
-
-interface AuthApi {
-    @GET("/api/users/refresh")
-    fun refreshToken(
-        @Header("Authorization") authorization: String
-    ): Call<RefreshTokenResponse>
-}
-
-interface VoteApi {
-    @Multipart
-    @POST("api/votes/create")
-    fun createVoteWithImage(
-        @Part("create_vote_json") createVoteJson: RequestBody,
-        @Part images: List<MultipartBody.Part>
-    ): Call<CreateVoteResponse>
-
-    @Multipart
-    @POST("api/votes/create")
-    fun createVoteWithoutImage(
-        @Part("create_vote_json") createVoteJson: RequestBody
-    ): Call<CreateVoteResponse>
-
-    @GET("/api/votes/list")
-    fun getVotes(
-        @Query("start_cursor") startCursor: String? = null,
-        @Query("category") category: String? = null,
-        @Header("Authorization") token: String
-    ): Call<VoteListResponse>
-
-    @GET("/api/votes/{vote_id}")
-    fun getVoteDetails(
-        @Path("vote_id") voteId: Int,
-        @Header("Authorization") authToken: String
-    ): Call<VoteDetailResponse>
-
-    @POST("/api/votes/{vote_id}/participate")
-    fun participateInVote(
-        @Path("vote_id") voteId: Int,
-        @Header("Authorization") authorization: String,
-        @Body requestBody: ParticipationRequest
-    ): Call<VoteDetailResponse>
-
-    @POST("api/votes/{voteId}/comment")
-    fun postComment(
-        @Path("voteId") voteId: Int,
-        @Header("Authorization") token: String,
-        @Body commentRequest: CommentRequest
-    ): Call<VoteDetailResponse>
-
-    @PATCH("api/votes/{vote_id}/comment/{comment_id}")
-    fun updateComment(
-        @Path("vote_id") voteId: Int,
-        @Path("comment_id") commentId: Int,
-        @Header("Authorization") token: String,
-        @Body request: CommentRequest
-    ): Call<VoteDetailResponse>
-
-    @DELETE("api/votes/{vote_id}/comment/{comment_id}")
-    fun deleteComment(
-        @Path("vote_id") voteId: Int,
-        @Path("comment_id") commentId: Int,
-        @Header("Authorization") token: String
-    ): Call<VoteDetailResponse>
-}
 
 // 요청 데이터 클래스
 data class SignupRequest(
@@ -302,6 +205,3 @@ data class GetMeResponse(
     val email: String,
     val college: Int
 )
-
-
-
