+++ app/src/main/java/com/example/wafflestudio_toyproject/network/AuthInterceptor.kt
@@ -1,0 +1,20 @@
+package com.example.wafflestudio_toyproject.network
+
+import android.content.SharedPreferences
+import okhttp3.Interceptor
+import okhttp3.OkHttpClient
+import okhttp3.Request
+import okhttp3.Response
+
+class AuthInterceptor(private val sharedPreferences: SharedPreferences) : Interceptor {
+    override fun intercept(chain: Interceptor.Chain): Response {
+        val accessToken = sharedPreferences.getString("access_token", null)
+        val requestBuilder: Request.Builder = chain.request().newBuilder()
+
+        if (!accessToken.isNullOrEmpty()) {
+            requestBuilder.addHeader("Authorization", "Bearer $accessToken")
+        }
+
+        return chain.proceed(requestBuilder.build())
+    }
+}
