+++ app/src/main/java/com/example/wafflestudio_toyproject/network/NetworkModule.kt
@@ -2,9 +2,6 @@ package com.example.wafflestudio_toyproject.network
 
 import android.content.Context
 import android.content.SharedPreferences
-import com.example.wafflestudio_toyproject.AuthApi
-import com.example.wafflestudio_toyproject.UserApi
-import com.example.wafflestudio_toyproject.VoteApi
 import dagger.Module
 import dagger.Provides
 import dagger.hilt.InstallIn
@@ -23,31 +20,28 @@ object NetworkModule {
 
     private const val BASE_URL = "http://52.78.27.95/"
 
-    // 일반 API용 OkHttpClient (Authenticator 포함)
+    @Provides
+    @Singleton
+    fun provideAuthInterceptor(sharedPreferences: SharedPreferences): AuthInterceptor {
+        return AuthInterceptor(sharedPreferences)
+    }
+
+    // 일반 API용 OkHttpClient (AuthInterceptor 추가)
     @Provides
     @Singleton
     fun provideOkHttpClient(
-        sharedPreferences: SharedPreferences,
+        authInterceptor: AuthInterceptor,
         tokenAuthenticator: TokenAuthenticator,
         loggingInterceptor: HttpLoggingInterceptor
     ): OkHttpClient {
         return OkHttpClient.Builder()
             .authenticator(tokenAuthenticator) // Authenticator 추가
-            .addInterceptor { chain ->
-                val requestBuilder = chain.request().newBuilder()
-                val accessToken = sharedPreferences.getString("access_token", null)
-
-                if (!accessToken.isNullOrEmpty()) {
-                    requestBuilder.addHeader("Authorization", "Bearer $accessToken")
-                }
-
-                chain.proceed(requestBuilder.build())
-            }
+            .addInterceptor(authInterceptor) // AuthInterceptor 추가
             .addInterceptor(loggingInterceptor)
             .build()
     }
 
-    // AuthAPI 전용 OkHttpClient (Authenticator 미포함)
+    // AuthAPI 전용 OkHttpClient (Authorization 불필요)
     @Provides
     @Singleton
     @Named("AuthClient")
@@ -59,7 +53,6 @@ object NetworkModule {
             .build()
     }
 
-    // 일반 API용 Retrofit (OkHttpClient 사용)
     @Provides
     @Singleton
     fun provideRetrofit(okHttpClient: OkHttpClient): Retrofit {
@@ -70,7 +63,6 @@ object NetworkModule {
             .build()
     }
 
-    // AuthAPI 전용 Retrofit (AuthClient 사용)
     @Provides
     @Singleton
     @Named("AuthRetrofit")
@@ -82,14 +74,12 @@ object NetworkModule {
             .build()
     }
 
-    // 일반 API UserApi 생성
     @Provides
     @Singleton
     fun provideUserApi(retrofit: Retrofit): UserApi {
         return retrofit.create(UserApi::class.java)
     }
 
-    // AuthAPI 생성 (AuthRetrofit 사용)
     @Provides
     @Singleton
     @Named("AuthApi")
@@ -97,14 +87,12 @@ object NetworkModule {
         return authRetrofit.create(AuthApi::class.java)
     }
 
-    // SharedPreferences 제공
     @Provides
     @Singleton
     fun provideSharedPreferences(@ApplicationContext context: Context): SharedPreferences {
         return context.getSharedPreferences("auth_prefs", Context.MODE_PRIVATE)
     }
 
-    //투표 api
     @Provides
     @Singleton
     fun provideVoteApi(retrofit: Retrofit): VoteApi {
@@ -115,7 +103,7 @@ object NetworkModule {
     @Singleton
     fun provideHttpLoggingInterceptor(): HttpLoggingInterceptor {
         return HttpLoggingInterceptor().apply {
-            level = HttpLoggingInterceptor.Level.BODY // FULL 요청/응답 로깅
+            level = HttpLoggingInterceptor.Level.BODY
         }
     }
 }
