+++ app/src/main/res/layout/activity_signup.xml
@@ -121,6 +121,7 @@
                     android:inputType="textPassword"
                     android:textColor="#000000"
                     android:textSize="16sp"
+                    android:hint="영문, 숫자, 특수문자 2종류 이상 조합된 8~20자"
                     android:backgroundTint="@android:color/transparent"/>
             </com.google.android.material.textfield.TextInputLayout>
